requirement.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
numpy==1.21.0
oauthlib==3.1.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pyxlsb==1.0.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.16.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
xlrd==2.0.1
xlwings==0.23.4
